rwuio: fix 32 bit wide access on 64 bit platforms
"long" and "unsigned long" are architecture specific, i.e. 4 bytes on
32bits platforms, and 8 bytes on 64bits platforms. "unsigned int" is
always 4 bytes.

Access width is very important for hardware register access as it may
exercise different logic in the controller.

+++ tools/rwuio.c
@@ -137,7 +137,7 @@ static void uio_read(struct uio_info_t *info)
 			break;
 			case 4:
 				printf("%08lx: %08x\n", addr,
-				       *(volatile unsigned long*)(vp));
+				       *(volatile unsigned int*)(vp));
 			break;
 			case 2:
 				printf("%08lx: %04x\n", addr,
